requirements.txt
Django==1.10
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.11
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
Django==1.10
django-filter==1.1.0
djangorestframework==3.7.7
google-api-core==0.1.4
google-auth==1.4.1
google-cloud-core==0.28.0
google-cloud-storage==1.7.0
google-cloud-vision==0.30.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.9.1
idna==2.6
Markdown==2.6.11
Pillow==5.0.0
protobuf==3.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.2.0
python-pygaze==0.6.0a25
pytz==2018.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
